Removed the dependency on the Glob package.
The Glob package was providing a minor utility function for automatically running doctest over all the *.hs files in the project, but was also consistently causing compilation issues. The minor convenience of globbing all the *.hs files is not worth the compilation issues, so I got rid of the dependency.