Giving up on Travis (for now?)
Getting doctests to work reliably now that I have an external dependency has been way too much of a headache to be worth it. I'm just going to ignore them for now. Hoping that better doctest integration comes along to either haskell.nix or cabal new-style builds so that I don't have to deal with this any more.